requirements.txt
python-telegram-bot>=20.0
python-dotenv>=0.19.0
faster-whisper>=0.9.0
jdatetime>=4.1.0
pytz>=2023.3
schedule>=1.2.0
persian-tools>=0.0.13
numpy>=1.24.0
torch>=2.0.0
torchaudio>=2.0.0
huggingface-hub==0.21.4
google-cloud-speech>=2.21.0
google-cloud-translate>=3.12.0
google-cloud-language>=2.11.0 